+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Features
 
-- feat: add support for custom plugins in httpFileStore initialization (AnWeber/vscode-httpyac#186)
+- add support for custom plugins in httpFileStore initialization (AnWeber/vscode-httpyac#186)
+- add modifier `$input-askonce` to ask only once for $input (#436)
 
 ## 6.3.4 (2023-03-30)