+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 6.4.0 (2023-04-29)
 
 ### Features